refactor: the whole thing
