+++ package.json
@@ -5,11 +5,11 @@
   "main": "src/index.js",
   "browser": {
     "node-webcrypto-ossl": false,
-    "./src/crypto/webcrypto.js": "./src/crypto/webcrypto-browser.js",
-    "./src/crypto/hmac.js": "./src/crypto/hmac-browser.js",
-    "./src/crypto/ecdh.js": "./src/crypto/ecdh-browser.js",
-    "./src/crypto/ciphers.js": "./src/crypto/ciphers-browser.js",
-    "./src/crypto/rsa.js": "./src/crypto/rsa-browser.js"
+    "./src/webcrypto/index.js": "./src/webcrypto/index-browser.js",
+    "./src/hmac/index.js": "./src/hmac/index-browser.js",
+    "./src/keys/ecdh.js": "./src/keys/ecdh-browser.js",
+    "./src/aes/ciphers.js": "./src/aes/ciphers-browser.js",
+    "./src/keys/rsa.js": "./src/keys/rsa-browser.js"
   },
   "scripts": {
     "lint": "aegir-lint",
@@ -30,7 +30,7 @@
     "crypto",
     "rsa"
   ],
-  "author": "Friedel Ziegelmayer <dignifiedqurie@gmail.com>",
+  "author": "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
   "license": "MIT",
   "dependencies": {
     "asn1.js": "^4.9.1",
