--- src/crypto.js
@@ -1,8 +1,0 @@
-'use strict'
-
-exports.webcrypto = require('./crypto/webcrypto')()
-exports.hmac = require('./crypto/hmac')
-exports.ecdh = require('./crypto/ecdh')
-exports.aes = require('./crypto/aes')
-exports.rsa = require('./crypto/rsa')
-exports.ed25519 = require('./crypto/ed25519')
