+++ src/hmac/lengths.js
