+++ src/keys/ephemeral-keys.js
@@ -1,11 +1,11 @@
 'use strict'
 
-const crypto = require('./crypto')
+const ecdh = require('./ecdh')
 
 // Generates an ephemeral public key and returns a function that will compute
 // the shared secret key.
 //
 // Focuses only on ECDH now, but can be made more general in the future.
 module.exports = (curve, callback) => {
-  crypto.ecdh.generateEphmeralKeyPair(curve, callback)
+  ecdh.generateEphmeralKeyPair(curve, callback)
 }
