+++ test/keys/rsa.spec.js
@@ -7,26 +7,25 @@ const expect = chai.expect
 chai.use(dirtyChai)
 const Buffer = require('safe-buffer').Buffer
 
-const crypto = require('../src')
-const rsa = crypto.keys.rsa
-const fixtures = require('./fixtures/go-key-rsa')
+const crypto = require('../../src')
+const rsa = crypto.keys.keys.rsa
+const fixtures = require('../fixtures/go-key-rsa')
 
 describe('RSA', () => {
   let key
+
   before((done) => {
-    crypto.generateKeyPair('RSA', 2048, (err, _key) => {
-      if (err) return done(err)
+    crypto.keys.generateKeyPair('RSA', 2048, (err, _key) => {
+      if (err) {
+        return done(err)
+      }
       key = _key
       done()
     })
   })
 
   it('generates a valid key', (done) => {
-    expect(
-      key
-    ).to.be.an.instanceof(
-      rsa.RsaPrivateKey
-    )
+    expect(key).to.be.an.instanceof(rsa.RsaPrivateKey)
 
     key.hash((err, digest) => {
       if (err) {
@@ -65,68 +64,35 @@ describe('RSA', () => {
       }
       const keyMarshal2 = key2.marshal()
 
-      expect(
-        keyMarshal
-      ).to.be.eql(
-        keyMarshal2
-      )
+      expect(keyMarshal).to.eql(keyMarshal2)
 
       const pk = key.public
       const pkMarshal = pk.marshal()
       const pk2 = rsa.unmarshalRsaPublicKey(pkMarshal)
       const pkMarshal2 = pk2.marshal()
 
-      expect(
-        pkMarshal
-      ).to.be.eql(
-        pkMarshal2
-      )
+      expect(pkMarshal).to.eql(pkMarshal2)
       done()
     })
   })
 
   describe('key equals', () => {
     it('equals itself', () => {
-      expect(
-        key.equals(key)
-      ).to.be.eql(
-        true
-      )
+      expect(key.equals(key)).to.eql(true)
 
-      expect(
-        key.public.equals(key.public)
-      ).to.be.eql(
-        true
-      )
+      expect(key.public.equals(key.public)).to.eql(true)
     })
 
     it('not equals other key', (done) => {
-      crypto.generateKeyPair('RSA', 2048, (err, key2) => {
-        if (err) return done(err)
+      crypto.keys.generateKeyPair('RSA', 2048, (err, key2) => {
+        if (err) {
+          return done(err)
+        }
 
-        expect(
-          key.equals(key2)
-        ).to.be.eql(
-          false
-        )
-
-        expect(
-          key2.equals(key)
-        ).to.be.eql(
-          false
-        )
-
-        expect(
-          key.public.equals(key2.public)
-        ).to.be.eql(
-          false
-        )
-
-        expect(
-          key2.public.equals(key.public)
-        ).to.be.eql(
-          false
-        )
+        expect(key.equals(key2)).to.eql(false)
+        expect(key2.equals(key)).to.eql(false)
+        expect(key.public.equals(key2.public)).to.eql(false)
+        expect(key2.public.equals(key.public)).to.eql(false)
         done()
       })
     })
@@ -168,7 +134,7 @@ describe('RSA', () => {
 
   describe('go interop', () => {
     it('verifies with data from go', (done) => {
-      const key = crypto.unmarshalPublicKey(fixtures.verify.publicKey)
+      const key = crypto.keys.unmarshalPublicKey(fixtures.verify.publicKey)
 
       key.verify(fixtures.verify.data, fixtures.verify.signature, (err, ok) => {
         if (err) throw err
